% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccr-package.R
\docType{package}
\name{bccr-package}
\alias{bccr}
\alias{bccr-package}
\title{bccr: cliente no oficial para la API de indicadores economicos del BCCR}
\description{
Funciones principales:
\itemize{
\item \code{\link[=descargar_bccr]{descargar_bccr()}}: descarga varios indicadores en una sola tabla.
\item \code{\link[=get_serie_bccr]{get_serie_bccr()}}: descarga un indicador.
\item \code{\link[=guardar_bccr]{guardar_bccr()}} / \code{\link[=leer_bccr]{leer_bccr()}}: guardan y leen los datos en parquet.
}
}
\details{
La API key se lee de la variable de entorno \code{BCCR_API_KEY}. Configurala
una vez por maquina con \code{usethis::edit_r_environ()}, agregando la linea
\verb{BCCR_API_KEY=<tu key>}, y reinicia R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fabianbt6/bccr-api}
  \item Report bugs at \url{https://github.com/fabianbt6/bccr-api/issues}
}

}
\author{
\strong{Maintainer}: Fabián Brenes \email{fabianbt6@gmail.com}

Authors:
\itemize{
  \item Fabián Brenes \email{fabianbt6@gmail.com}
}

}
\keyword{internal}
