% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{indicadores_default}
\alias{indicadores_default}
\title{Indicadores por defecto}
\usage{
indicadores_default()
}
\value{
Vector de caracteres con codigos de indicadores.
}
\description{
Lista de codigos de indicadores del BCCR que se descargan si no se
indica otra cosa en \code{\link[=descargar_bccr]{descargar_bccr()}}.
}
\examples{
indicadores_default()
}
